Add: コンポーネントの切り分け
