+++ src/components/Styles.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const SButton = styled.button`
+  padding: 4px 8px;
+  background-color: lightblue;
+  border: none;
+  border-radius: 2px;
+  &:hover {
+    background-color: #4b7b8b;
+    color: white;
+    cursor: pointer;
+  }
+`;
